# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Management API",
        "description": "API documentation for Event Management System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/panel/events": {
            "get": {
                "tags": [
                    "Panel - Events"
                ],
                "summary": "Get all events (Admin)",
                "description": "Retrieve paginated list of all events for admin panel",
                "operationId": "6310392b119afc0e87e3c663d0e89421",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of events",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventPanelListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin access required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Panel - Events"
                ],
                "summary": "Create a new event (Admin)",
                "description": "Create a new event",
                "operationId": "82868e644f14fb55fdf81b3e3cf57f1e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EventRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Event created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventPanelDetailResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin access required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/panel/events/{event}": {
            "get": {
                "tags": [
                    "Panel - Events"
                ],
                "summary": "Get event details (Admin)",
                "description": "Retrieve detailed information about a specific event",
                "operationId": "cc43b8453bc012549923100a71428297",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventPanelDetailResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin access required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Panel - Events"
                ],
                "summary": "Update an event (Admin)",
                "description": "Update an existing event",
                "operationId": "d8f131dc24b2c66599dc2b9b2cdc3379",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EventRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Event updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventPanelDetailResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin access required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Panel - Events"
                ],
                "summary": "Delete an event (Admin)",
                "description": "Delete an event",
                "operationId": "fbfe93d2efa3fec8549e95beb677bbaa",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin access required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EventPanelListResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "data": {
                        "properties": {
                            "events": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Et autem et rem."
                                        },
                                        "capacity": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "free_capacity": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "creator": {
                                            "type": "string",
                                            "example": "Admin User"
                                        },
                                        "active_reserv": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2026-02-20T14:57:33.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "page_data": {
                                "properties": {
                                    "current_page": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "per_page": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "total": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "last_page": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "from": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "to": {
                                        "type": "integer",
                                        "example": 10,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "EventPanelDetailResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "data": {
                        "properties": {
                            "event": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Event name"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "free_capacity": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "creator": {
                                        "type": "string",
                                        "example": "Admin User"
                                    },
                                    "active_reserv": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "created_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-01-01T10:00:00.000000Z"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "EventRequest": {
                "required": [
                    "title",
                    "capacity"
                ],
                "properties": {
                    "title": {
                        "description": "Event title (unique)",
                        "type": "string",
                        "maxLength": 80,
                        "example": "Event Title"
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 250,
                        "example": "Event description",
                        "nullable": true
                    },
                    "capacity": {
                        "description": "Event capacity",
                        "type": "integer",
                        "minimum": 1,
                        "example": 100
                    },
                    "start_date": {
                        "description": "Event start date (Y-m-d H:i)",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01 10:00",
                        "nullable": true
                    },
                    "end_date": {
                        "description": "Event end date (Y-m-d H:i, must be after start_date)",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01 18:00",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "SuccessResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Error message"
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "UnauthenticatedResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "ForbiddenResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Forbidden"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your access token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}